refactor: move utilities to global utils
