+++ packages/ui/client/composables/client/index.ts
@@ -13,8 +13,7 @@ import { computed, reactive as reactiveVue, ref, shallowRef, watch } from 'vue'
 import { explorerTree } from '~/composables/explorer'
 import { isFileNode } from '~/composables/explorer/utils'
 import { isSuite as isTaskSuite } from '~/utils/task'
-import { createFileTask } from '../../../../vitest/src/runtime/runner/utils/collect'
-import { getTasks } from '../../../../vitest/src/runtime/runner/utils/tasks'
+import { createFileTask, getTasks } from '../../../../vitest/src/utils/tasks'
 import { ui } from '../../composables/api'
 import { ENTRY_URL, isReport } from '../../constants'
 import { parseError } from '../error'
