+++ packages/vitest/src/node/core.ts
@@ -25,13 +25,11 @@ import { join, normalize, relative } from 'pathe'
 import { isRunnableDevEnvironment } from 'vite'
 import { version } from '../../package.json' with { type: 'json' }
 import { distDir } from '../paths'
-import { interpretTaskModes, someTasksAreOnly } from '../runtime/runner/utils/collect'
 import { createTagsFilter } from '../runtime/runner/utils/tags'
-import { getTasks, hasFailed } from '../runtime/runner/utils/tasks'
 import { wildcardPatternToRegExp } from '../utils/base'
 import { limitConcurrency } from '../utils/limit-concurrency'
 import { NativeModuleRunner } from '../utils/nativeModuleRunner'
-import { convertTasksToEvents } from '../utils/tasks'
+import { convertTasksToEvents, getTasks, hasFailed, interpretTaskModes, someTasksAreOnly } from '../utils/tasks'
 import { Traces } from '../utils/traces'
 import { astCollectTests, createFailedFileTask } from './ast-collect'
 import { BrowserSessions } from './browser/sessions'
