+++ packages/vitest/src/node/pools/workers/typecheckWorker.ts
@@ -7,8 +7,8 @@ import type { TestRunEndReason } from '../../types/reporter'
 import type { PoolOptions, PoolWorker, WorkerRequest, WorkerResponse } from '../types'
 import EventEmitter from 'node:events'
 import { createDefer } from '@vitest/utils/helpers'
-import { hasFailed } from '../../../runtime/runner/utils/tasks'
 import { Typechecker } from '../../../typecheck/typechecker'
+import { hasFailed } from '../../../utils/tasks'
 
 /** @experimental */
 export class TypecheckPoolWorker implements PoolWorker {
