+++ packages/vitest/src/node/reporters/base.ts
@@ -11,10 +11,9 @@ import { toArray } from '@vitest/utils/helpers'
 import { parseStacktrace } from '@vitest/utils/source-map'
 import { relative } from 'pathe'
 import c from 'tinyrainbow'
-import { getSuites, getTestName, getTests, hasFailed } from '../../runtime/runner/utils/tasks'
 import { groupBy } from '../../utils/base'
 import { isTTY } from '../../utils/env'
-import { hasFailedSnapshot } from '../../utils/tasks'
+import { getSuites, getTestName, getTests, hasFailed, hasFailedSnapshot } from '../../utils/tasks'
 import { generateCodeFrame, printStack } from '../printError'
 import { BENCH_TABLE_HEAD, computeBenchColumnWidths, padBenchRow, renderBenchmarkRow } from './renderers/benchmark-table'
 import { F_CHECK, F_DOWN_RIGHT, F_POINTER } from './renderers/figures'
