+++ packages/vitest/src/node/reporters/junit.ts
@@ -9,8 +9,8 @@ import { existsSync, promises as fs } from 'node:fs'
 import { hostname } from 'node:os'
 import { stripVTControlCharacters } from 'node:util'
 import { basename, dirname, relative, resolve } from 'pathe'
-import { getSuites } from '../../runtime/runner/utils/tasks'
 import { getOutputFile } from '../../utils/config-helpers'
+import { getSuites } from '../../utils/tasks'
 import { renderBenchmarkTableText } from './renderers/benchmark-table'
 import { IndentedLogger } from './renderers/indented-logger'
