+++ packages/vitest/src/node/reporters/verbose.ts
@@ -1,6 +1,6 @@
 import type { TestCase, TestModule } from './reported-tasks'
 import c from 'tinyrainbow'
-import { getTestName } from '../../runtime/runner/utils/tasks'
+import { getTestName } from '../../utils/tasks'
 import { DefaultReporter } from './default'
 import { F_RIGHT } from './renderers/figures'
 import { separator } from './renderers/utils'
