+++ packages/vitest/src/runtime/runners/test.ts
@@ -24,15 +24,14 @@ import { createExpect } from '../../integrations/chai/index'
 import { inject } from '../../integrations/inject'
 import { getSnapshotClient } from '../../integrations/snapshot/chai'
 import { vi } from '../../integrations/vi'
+import { createFileTask, getNames, getTestName, getTests } from '../../utils/tasks'
 import { createBench, kFinalize } from '../benchmark'
 import { rpc } from '../rpc'
 import { getFn, getHooks } from '../runner/map'
 import { createTaskCollector, getCurrentSuite } from '../runner/suite'
 import { getCurrentTest } from '../runner/test-state'
 import { createChainable } from '../runner/utils/chain'
-import { createFileTask } from '../runner/utils/collect'
 import { matchesTags } from '../runner/utils/tags'
-import { getNames, getTestName, getTests } from '../runner/utils/tasks'
 import { getWorkerState } from '../utils'
 
 export class TestRunner implements VitestTestRunner {
